+++ src/main/java/io/alexdo/mixtech/jpa/entity/SongEntity.java
@@ -26,7 +26,7 @@ public class SongEntity {
     private Float liveness;
     private Float valence;
     private Float tempo;
-    private Long durationMs;
+    private Integer durationMs;
     private Integer timeSignature;
     private Integer popularity;
 }
